+++ template/tests/hello_test.rs
@@ -1,4 +1,4 @@
-//INCLUDEFILE !embassy tests/hello_test.rs
+//INCLUDEFILE !option("embassy") && option("embedded-test")
 //! Demo test suite using embedded-test
 //!
 //! You can run this using `cargo test` as usual.