+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- The generated project no longer enables `static_cell/nightly` (#199)
+
 ### Fixed
 
 ### Removed