+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added `rust-version` to the generated Cargo.toml (#192)
+- Generate settings for Zed (#200)
 
 ### Changed