+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Test source is no longer generated if not needed (#201)
+
 ### Removed
 
 ## [0.4.0] - 2025-06-06